Add torch sim Interface

+++ nequix/torch_impl/model.py
@@ -558,7 +558,7 @@ class NequixTorch(torch.nn.Module):
         receivers: torch.Tensor,
     ):
         # input features are one-hot encoded species
-        features = torch.nn.functional.one_hot(species, self.n_species).to(torch.float32)
+        features = torch.nn.functional.one_hot(species, self.n_species).to(displacements.dtype)
         r_norm = torch.linalg.norm(displacements, ord=2, dim=-1)
 
         radial_basis = (
